Progress on regridding in parallel (distributed MPI)

+++ contrib/bamg/include/InterpFromMeshToMesh2dCavities.h
@@ -15,18 +15,18 @@ typedef unsigned long ulong;
 
 /*threading: */
 typedef struct{
-	ulong* 	PreviousNumbering;
-	ulong* 	old_elements;
-	ulong* 	new_elements;
-	ulong* 	size_born_cavity;
-	ulong* 	size_dead_cavity;
-	ulong* 	element_born_cavity;
-	ulong* 	element_dead_cavity;
-	ulong 	new_bamg_mesh_Nn;
-	ulong 	nb_matching_elements;
-	ulong   nb_cavities;
-	ulong	max_size_born_cavity;
-    ulong	max_size_dead_cavity;
+	double* PreviousNumbering;
+	double* old_elements;
+	double* new_elements;
+	int* 	size_born_cavity;
+	int* 	size_dead_cavity;
+	double* element_born_cavity;
+	double* element_dead_cavity;
+	int 	new_bamg_mesh_Nn;
+	int 	nb_matching_elements;
+	int     nb_cavities;
+	int  	max_size_born_cavity;
+    int  	max_size_dead_cavity;
 } InterpFromMeshToMesh2dCavitiesThreadStruct;
 
 #define next(a) ( ( (a) == 0 )  ?  1   : ( ( (a) == 1 ) ?  2   : 0 ) )
@@ -37,9 +37,9 @@ int InterpFromMeshToMesh2dCavities(double** pdata_interp,double* data,int N_data
 
 int DetectCavities(InterpFromMeshToMesh2dCavitiesThreadStruct* gate, BamgMesh* bamgmesh_old,BamgMesh* bamgmesh_new);
 
-int InterpCavity(double* tmp_mean_variables, double* tmp_integrated_area, 
-			int nb_dead_elements, int nb_born_elements, int nb_variables, 
-			ulong* dead_elements, ulong* born_elements, ulong* PreviousNumbering, 
+int InterpCavity(double* tmp_mean_variables, double* tmp_integrated_area,
+			int nb_dead_elements, int nb_born_elements, int nb_variables,
+			double* dead_elements, double* born_elements, double* PreviousNumbering,
 			double* IntMatrix_in, BamgMesh* bamgmesh_old, BamgMesh* bamgmesh_new, int debug_born_elements_i);
 
 #endif
